refactor: indent extension 로직 분리
